add onClick for modal, fixed SelectionRequest component, add logo to modal
