+++ src/components/features/HomePage/Main/blockPlatform/BlockPlatform.tsx
@@ -7,8 +7,12 @@ import { Button } from "@/src/components/shared/buttons/Button";
 import Slider from "@/src/components/shared/slider/Slider";
 import ListCardsPlatforms from "@/src/components/entities/lists/listCardsPlatforms/ListCardsPlatforms";
 import { CARDS_PLATFORMS } from "@/src/components/shared/slider/CardsPlatformsConfig";
+import { useModal } from "@/src/hooks/useModal";
+import Modal from "@/src/components/shared/modal/Modal";
+import SelectionRequest from "@/src/components/entities/selectionRequest/SelectionRequest";
 
 const BlockPlatform = () => {
+     const { isShown, toggle } = useModal();
      return (
           <div className={styles.wrapper}>
                <div className={styles.blockText}>
@@ -20,7 +24,7 @@ const BlockPlatform = () => {
                          бизнес-задачу
                     </Text>
                     <div className={styles.buttons}>
-                         <Button type="button" active={true} width={250}>
+                         <Button type="button" active={true} width={250} onClick={toggle}>
                               Подобрать платформу
                          </Button>
                          <LinkShowAllCards href="/" />
@@ -29,6 +33,9 @@ const BlockPlatform = () => {
                <Slider>
                     <ListCardsPlatforms config={CARDS_PLATFORMS} />
                </Slider>
+               <Modal isShown={isShown} hide={toggle}>
+                    <SelectionRequest close={toggle} />
+               </Modal>
           </div>
      );
 };
