+++ src/components/features/HomePage/Main/blockSelectionBot/BlockSelectionBot.tsx
@@ -19,46 +19,13 @@ interface IUserRequest {
 export const BlockSelectionBot = () => {
      const { isShown, toggle } = useModal();
 
-     const selectionRequest: Array<IUserRequest> = [
-          {
-               htmlFor: "first_name",
-               label: "Имя",
-               type: "text",
-               name: "first_name",
-               placeholder: "Имя",
-          },
-          {
-               htmlFor: "email",
-               label: "E-mail",
-               type: "email",
-               name: "email",
-               placeholder: "example@mail.com",
-          },
-          {
-               htmlFor: "telephone",
-               label: "Телефон",
-               type: "tel",
-               name: "tel",
-               placeholder: "+7",
-          },
-          {
-               htmlFor: "comment",
-               label: "Комментарий",
-               type: "text",
-               name: "comment",
-               placeholder: "Текст (200 символов)",
-          },
-     ];
-
      return (
           <div className={css.wrapper}>
                <div className={css.container}>
                     <div className={css.blockLeft}>
                          <div className={css.textInfo}>
                               <Title type="h1" color="black">
-                                   Подбери{" "}
-                                   <ElemChooseChatBot variant={ElemVariantProps.home}/>
-                                   и автоматизируй свои продажи
+                                   Подбери <ElemChooseChatBot variant={ElemVariantProps.home} />и автоматизируй свои продажи
                               </Title>
 
                               <Text type="reg24" color="grey">
@@ -67,7 +34,7 @@ export const BlockSelectionBot = () => {
                          </div>
 
                          <div className={css.button}>
-                              <Button type="button" active={true}  onClick={toggle}>
+                              <Button type="button" active={true} onClick={toggle}>
                                    Подобрать решение
                               </Button>
                          </div>
@@ -93,7 +60,7 @@ export const BlockSelectionBot = () => {
                     </div>
                </div>
                <Modal isShown={isShown} hide={toggle}>
-                    <SelectionRequest schema={selectionRequest} close={toggle} />
+                    <SelectionRequest close={toggle} />
                </Modal>
           </div>
      );
